Add method to validate table index in TableAPI class.
